# api.yaml
openapi: 3.0.0
info:
  title: Dispatcher API
  description: HTTP API for Dispatcher service
  version: 0.0.2
paths:
  /nodes/{node_name}:
    post:
      summary: Update network node configuration
      operationId: post_node
      parameters: 
      - name: node_name
        in: path
        description: Node name
        required: true
        schema:
          type: string
          format: "s%d"
      requestBody:
        description: Optional description in *Markdown*
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeConfig'
      tags:
        - mininet
      responses:
        200:
          description: Successful operation
  /switch/{switch_name}:
    post:
      summary: Create new switch in the network
      operationId: addSwitch
      parameters: 
      - name: switch_name
        in: path
        description: Switch name
        required: true
        schema:
          type: string
          format: "s%d"

      tags:
        - mininet
      responses:
        200:
          description: Succesful operation
        403:
          description: Switch already exists
  /link/:
    post:
      summary: Create new link between nodes
      operationId: add_link
      tags:
        - mininet
      parameters:
        - $ref: "#/components/parameters/FromSwitchId"
        - $ref: "#/components/parameters/ToSwitchId"
      responses:
        200:
          description: Succesful operation
        403:
          description: Node(s) does not exist
...
  /nodes/{node_name}/cmd:
    post:
      summary: Run command for node specified
      operationId: do_cmd
      parameters: 
      - name: node_name
        in: path
        description: Node name
        required: true
        schema:
          type: string
          format: "s%d"
      tags:
        - mininet
      requestBody:
        description: Optional description in *Markdown*
        required: true
        content:
          text/plain:
            schema:
              type: string  
      responses:
        200:
          description: Succesful operation
...
components:
  parameters:
    FromSwitchId:
      in: query
      name: from_switch
      schema:
        type: string
      required: true
    ToSwitchId:
      in: query
      name: to_switch
      schema:
        type: string
      required: true
...
  schemas:
...
    NodeConfig:
      type: object
      required:
        - params
      properties:
        params:
          $ref: "#/components/schemas/NodeParameter"
    NodeParameter:
      type: object
      properties:
        delay:
          type: string
          format: "%dms"
        bw:
          type: integer
...